Post facility Data working

+++ src/app/serializers.py
@@ -59,19 +59,22 @@ class GroupSerializer(serializers.ModelSerializer):
     person_set = PersonSerializer(many=True,read_only=True)
 
     def create(self, validated_data):
-        people_data = validated_data.pop("person_set")
-        group = Group.objects.create(**validated_data)
-        for person_data in people_data:
-            person_data['group'] = group.id
-            person_serializer = PersonSerializer(data=person_data)
-            if person_serializer.is_valid():
-                person = person_serializer.save()
-            else:
-                group.delete()
-                raise ValidationError(person_serializer.errors)
-        return group 
-   
-    
+        if "person_set" in validated_data.keys():
+            people_data = validated_data.pop("person_set")
+            group = Group.objects.create(**validated_data)
+            for person_data in people_data:
+                person_data['group'] = group.id
+                person_serializer = PersonSerializer(data=person_data)
+                if person_serializer.is_valid():
+                    person = person_serializer.save()
+                else:
+                    group.delete()
+                    raise ValidationError(person_serializer.errors)
+            return group 
+        else:
+            return super().create(validated_data)
+
+  
     class Meta():
         model = Group
         fields = ['id', 'category', 'count', 'facility_preference', 'person_set']
